+++ core/templates/dev/head/pages/exploration_editor/editor_tab/editor_tab.html
@@ -1,6 +1,6 @@
 <div class="row">
   <div class="col-lg-8 col-md-8 col-sm-8">
-    {% include 'pages/exploration_editor/editor_tab/exploration_editor_tab.html' %}
+    <exploration-editor-tab></exploration-editor-tab>
   </div>
   <div class="col-lg-4 col-md-4 col-sm-4">
     <div class="oppia-editor-sidebar hidden-xs hidden-sm">
